Reorganise code.
